requirements.txt
Flask==2.0.1
pandas==1.2.5
numpy==1.19.5
tensorflow-cpu
gunicorn